+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased](https://github.com/rerun-io/rerun/compare/latest...HEAD)
 * Add support for invisible tiles
-* `PartialEq` for `Tiles` now ignore internal state
+* `PartialEq` for `Tiles` now ignores internal state
 * Add `Tiles::find_pane`
 * Add `Tiles::remove_recursively`